refactor: column and task operation class names
