+++ src/main/java/dev/farneser/tasktracker/api/operations/commands/column/delete/DeleteColumnCommand.java
@@ -1,0 +1,14 @@
+package dev.farneser.tasktracker.api.operations.commands.column.delete;
+
+import dev.farneser.tasktracker.api.mediator.Command;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class DeleteColumnCommand implements Command<Void> {
+    private Long userId;
+    private Long columnId;
+}
