+++ src/main/java/dev/farneser/tasktracker/api/operations/commands/column/patch/PatchColumnCommand.java
@@ -1,0 +1,17 @@
+package dev.farneser.tasktracker.api.operations.commands.column.patch;
+
+import dev.farneser.tasktracker.api.mediator.Command;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class PatchColumnCommand implements Command<Void> {
+    private Long userId;
+    private Long columnId;
+    private String columnName;
+    private Boolean isCompleted;
+    private Long orderNumber;
+}
