--- src/main/java/dev/farneser/tasktracker/api/operations/commands/kanbancolumn/delete/DeleteKanbanColumnCommand.java
@@ -1,14 +1,0 @@
-package dev.farneser.tasktracker.api.operations.commands.kanbancolumn.delete;
-
-import dev.farneser.tasktracker.api.mediator.Command;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class DeleteKanbanColumnCommand implements Command<Void> {
-    private Long userId;
-    private Long columnId;
-}
