--- src/main/java/dev/farneser/tasktracker/api/operations/commands/kanbancolumn/patch/PatchKanbanColumnCommand.java
@@ -1,17 +1,0 @@
-package dev.farneser.tasktracker.api.operations.commands.kanbancolumn.patch;
-
-import dev.farneser.tasktracker.api.mediator.Command;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class PatchKanbanColumnCommand implements Command<Void> {
-    private Long userId;
-    private Long columnId;
-    private String columnName;
-    private Boolean isCompleted;
-    private Long orderNumber;
-}
