--- src/main/java/dev/farneser/tasktracker/api/operations/commands/kanbantask/delete/DeleteKanbanTaskCommand.java
@@ -1,14 +1,0 @@
-package dev.farneser.tasktracker.api.operations.commands.kanbantask.delete;
-
-import dev.farneser.tasktracker.api.mediator.Command;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class DeleteKanbanTaskCommand implements Command<Void> {
-    private Long userId;
-    private Long taskId;
-}
