--- src/main/java/dev/farneser/tasktracker/api/operations/queries/kanbancolumn/getbyid/GetKanbanColumnByIdQuery.java
@@ -1,13 +1,0 @@
-package dev.farneser.tasktracker.api.operations.queries.kanbancolumn.getbyid;
-
-import dev.farneser.tasktracker.api.mediator.Query;
-import dev.farneser.tasktracker.api.operations.views.KanbanColumnView;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class GetKanbanColumnByIdQuery implements Query<KanbanColumnView> {
-    private Long userId;
-    private Long columnId;
-}
